improve error messages of some http calls
